Refactor test.go: extract platform-specific code to separate files
Move iOS and Android driver functions to dedicated files for better organization:
- ios.go: iOS driver, simulator detection, device info, app installation
- android.go: Android driver, UIAutomator2 setup, device detection

Also removes unused WDADir constant and GetWDABasePath() from setup.go.
